Started Asset and Associations Database
This will modify the Show class so that it generates a separate json
file that keeps track of the shots and assets being associated to
each other; the Shot class will also update this json upon creation.

Also this will change some snake case formatting into camel case for
consistency.