Database Changes and Addition
This will remove the parameter for databasenames in the class
initializations as it was redundant and confusing. Database names will
simply be the input from the "name" parameter with "Database" following
it.

This will also add an Asset class.

An instance of Show will now create additional databases tracking
associations between the assets and shows in them.