Database Modification and Addition
This will get rid of databasename parameters as it was redundant and
confusing. The user will no longer need to input an argument for this.
Databases from class instances will be named after the value input for
"name" followed by "Database."

Also, this will automatically create databases in the Show directory
that track the associations between shows and assets. An Asset class
will also be added.