Associations Database Functionality
This will add a method to update the associations and category
database at the show directory; there will also be a new method that checks
whether or not a directory exists so that this can be used before
automatically populating the associations database. This is so that the
user can't accidently add associations for shows and assets that haven't
been created yet.

Instancing assets and shows will automatically update the associations and
category databases.