Name Parameter and Associate Method Addition
This will get rid of the name parameter from all classes. This parameter
caused confusion and potential for user error where the name of the instance
did not match directory name of that instance; the name of the instance will
now automatically be the name of its directory so they always match.

Rather than trying to automatically update the associations database
upon creation of an asset or shot instance, the user will be able to do
this after the instance is created and using the new associateAssetShot
method in the parent class. This will allow less required arguments for
the creation of an instance and easier maintenance of code.